🔀🎨 Refactor meldezeitraum (#81)
* :white_check_mark: updated validator test to only test the specific validator

* 🗃️ correct version number

* :art: rethrow all exceptions in validator as RuntimeExceptions

* :white_check_mark: shorten assert in test

* :safety_vest: added @notblank to match frontend

* 🚚 moved dtos in the respective package

* :art: snackbar message storing is part of service

* :safety_vest: dependent field validation changed

* :white_check_mark: frontend unit test fix

* :art: spotless fix

* :art: unused import

* :art: import order checkstyle
