+++ praktikumsplaner-backend/src/main/java/de/muenchen/oss/praktikumsplaner/domain/Meldezeitraum.java
@@ -2,6 +2,7 @@ package de.muenchen.oss.praktikumsplaner.domain;
 
 import de.muenchen.oss.praktikumsplaner.annotations.StartDateBeforeEndDate;
 import jakarta.persistence.Entity;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.time.LocalDate;
@@ -23,6 +24,7 @@ import lombok.ToString;
 public class Meldezeitraum extends BaseEntity {
     @NotNull
     @Size(max = 255, message = "Name für Zeitraum darf nicht länger als {max} Zeichen sein.")
+    @NotBlank
     private String zeitraumName;
 
     @NotNull
