+++ praktikumsplaner-backend/src/main/resources/db/migration/V2__Create_Meldezeitraum_Table.sql
