+++ praktikumsplaner-backend/src/test/java/de/muenchen/oss/praktikumsplaner/validators/StartDateBeforeEndDateValidatorTest.java
@@ -1,20 +1,23 @@
 package de.muenchen.oss.praktikumsplaner.validators;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import de.muenchen.oss.praktikumsplaner.domain.Meldezeitraum;
-import de.muenchen.oss.praktikumsplaner.dtos.CreateMeldezeitraumDTO;
-import de.muenchen.oss.praktikumsplaner.dtos.MeldezeitraumDTO;
+import de.muenchen.oss.praktikumsplaner.annotations.StartDateBeforeEndDate;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
 import jakarta.validation.ValidatorFactory;
 import java.time.LocalDate;
-import java.util.UUID;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 public class StartDateBeforeEndDateValidatorTest {
+
+    @StartDateBeforeEndDate(startDate = "startDate", endDate = "endDate")
+    private record TestClass(LocalDate startDate, LocalDate endDate) {
+    }
+
     private static Validator validator;
 
     @BeforeAll
@@ -28,25 +31,11 @@ public class StartDateBeforeEndDateValidatorTest {
         final LocalDate startDate = LocalDate.of(2020, 10, 10);
         final LocalDate endDate = LocalDate.of(2021, 10, 10);
 
-        final MeldezeitraumDTO meldezeitraumDTO = new MeldezeitraumDTO(
-                UUID.randomUUID(),
-                "Name",
+        final TestClass testObject = new TestClass(
                 startDate,
                 endDate);
 
-        final CreateMeldezeitraumDTO meldezeitraumCreateDTO = new CreateMeldezeitraumDTO(
-                "Name",
-                startDate,
-                endDate);
-
-        final Meldezeitraum meldezeitraum = new Meldezeitraum(
-                "Name",
-                startDate,
-                endDate);
-
-        assertTrue(validator.validate(meldezeitraumDTO).isEmpty());
-        assertTrue(validator.validate(meldezeitraumCreateDTO).isEmpty());
-        assertTrue(validator.validate(meldezeitraum).isEmpty());
+        assertTrue(validator.validate(testObject).isEmpty());
     }
 
     @Test
@@ -54,25 +43,11 @@ public class StartDateBeforeEndDateValidatorTest {
         final LocalDate startDate = LocalDate.of(2020, 10, 10);
         final LocalDate endDate = null;
 
-        final MeldezeitraumDTO meldezeitraumDTO = new MeldezeitraumDTO(
-                UUID.randomUUID(),
-                "Name",
+        final TestClass meldezeitraumDTO = new TestClass(
                 startDate,
                 endDate);
 
-        final CreateMeldezeitraumDTO meldezeitraumCreateDTO = new CreateMeldezeitraumDTO(
-                "Name",
-                startDate,
-                endDate);
-
-        final Meldezeitraum meldezeitraum = new Meldezeitraum(
-                "Name",
-                startDate,
-                endDate);
-
-        assertFalse(validator.validate(meldezeitraumDTO).isEmpty());
-        assertFalse(validator.validate(meldezeitraumCreateDTO).isEmpty());
-        assertFalse(validator.validate(meldezeitraum).isEmpty());
+        assertThrows(Exception.class, () -> validator.validate(meldezeitraumDTO));
     }
 
     @Test
@@ -80,24 +55,10 @@ public class StartDateBeforeEndDateValidatorTest {
         final LocalDate startDate = LocalDate.of(2020, 10, 10);
         final LocalDate endDate = LocalDate.of(2019, 10, 10);
 
-        final MeldezeitraumDTO meldezeitraumDTO = new MeldezeitraumDTO(
-                UUID.randomUUID(),
-                "Name",
-                startDate,
-                endDate);
-
-        final CreateMeldezeitraumDTO meldezeitraumCreateDTO = new CreateMeldezeitraumDTO(
-                "Name",
-                startDate,
-                endDate);
-
-        final Meldezeitraum meldezeitraum = new Meldezeitraum(
-                "Name",
+        final TestClass meldezeitraumDTO = new TestClass(
                 startDate,
                 endDate);
 
         assertFalse(validator.validate(meldezeitraumDTO).isEmpty());
-        assertFalse(validator.validate(meldezeitraumCreateDTO).isEmpty());
-        assertFalse(validator.validate(meldezeitraum).isEmpty());
     }
 }
