+++ praktikumsplaner-frontend/frontend/src/api/MeldezeitraumService.ts
@@ -1,6 +1,8 @@
 import FetchUtils from "@/api/FetchUtils";
 import { API_BASE, MELDEZEITRAUM_BASE } from "@/Constants";
 import Meldezeitraum from "@/types/Meldezeitraum";
+import {useSnackbarStore} from "@/stores/snackbar";
+import {Levels} from "@/api/error";
 
 export default class MeldezeitraumService {
     public create(meldezeitraum: Meldezeitraum): Promise<Meldezeitraum> {
@@ -9,10 +11,18 @@ export default class MeldezeitraumService {
             FetchUtils.getPOSTConfig(meldezeitraum)
         )
             .then((response) => {
+                useSnackbarStore().showMessage({
+                    message: "Meldezeitraum erfolgreich angelegt",
+                    level: Levels.INFO,
+                });
                 FetchUtils.defaultResponseHandler(response);
                 return response.json();
             })
             .catch((err) => {
+                useSnackbarStore().showMessage({
+                    message: err.message,
+                    level: Levels.ERROR,
+                });
                 FetchUtils.defaultResponseHandler(err);
             });
     }
