+++ praktikumsplaner-frontend/frontend/src/components/Meldezeitraeume/ZeitraumPicker.vue
@@ -8,7 +8,6 @@
                 :rules="startZeitpunktRules"
                 prepend-icon="mdi-calendar-start"
                 label="Startzeitpunkt"
-                @change="startZeitpunktChange"
             >
             </v-text-field>
         </v-col>
@@ -20,7 +19,6 @@
                 prepend-icon="mdi-calendar-end"
                 label="Endzeitpunkt"
                 :rules="endZeitpunktRules"
-                @change="endZeitpunktChange"
             >
             </v-text-field>
         </v-col>
@@ -30,40 +28,29 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { useRules } from "@/composables/rules";
-import Zeitraum from "@/types/Zeitraum";
+import Meldezeitraum from "@/types/Meldezeitraum";
 
 const props = defineProps<{
-    value: Zeitraum;
+    value: Meldezeitraum;
 }>();
 const emits = defineEmits<{
-    (e: "input", v: Zeitraum): void;
+    (e: "input", v: Meldezeitraum): void;
 }>();
 
 const validationRules = useRules();
 
-const range = computed(() => {
-    return props.value;
-});
-
-const startRef = computed(() => {
-    return props.value.startZeitpunkt;
-});
-const endRef = computed(() => {
-    return props.value.endZeitpunkt;
-});
+const range = computed(() => props.value);
 
 const isStartBeforeEnd = computed(() => {
-    if (endRef.value == undefined) return true;
     return (
-        (startRef.value && new Date(startRef.value) < new Date(endRef.value)) ||
+        range.value.isStartBeforeEnd ||
         "Das Startdatum muss vor dem Enddatum liegen."
     );
 });
 
 const isEndBeforeStart = computed(() => {
-    if (startRef.value == undefined) return true;
     return (
-        (endRef.value && new Date(startRef.value) < new Date(endRef.value)) ||
+        range.value.isStartBeforeEnd ||
         "Das Enddatum muss nach dem Startdatum liegen."
     );
 });
@@ -85,18 +72,4 @@ const startZeitpunktRules = computed(() => {
         isStartBeforeEnd.value,
     ];
 });
-
-function startZeitpunktChange() {
-    emits(
-        "input",
-        new Zeitraum(range.value.startZeitpunkt, range.value.endZeitpunkt)
-    );
-}
-
-function endZeitpunktChange() {
-    emits(
-        "input",
-        new Zeitraum(range.value.startZeitpunkt, range.value.endZeitpunkt)
-    );
-}
 </script>
