+++ praktikumsplaner-frontend/frontend/src/types/Meldezeitraum.ts
@@ -4,4 +4,12 @@ export default class Meldezeitraum {
         public startZeitpunkt?: string,
         public endZeitpunkt?: string
     ) {}
+
+    get isStartBeforeEnd(): boolean {
+        return (
+            !this.startZeitpunkt ||
+            !this.endZeitpunkt ||
+            new Date(this.startZeitpunkt) < new Date(this.endZeitpunkt)
+        );
+    }
 }
